[FEAT] : SearchClient 전략 인터페이스 및 SearchClientManager 유스케이스 구현
- SearchClient 인터페이스를 application.search.port 계층에 정의
- Naver/Kakao 클라이언트에서 해당 인터페이스 구현
- SearchClientManager에서 다중 SearchClient를 순차 fallback 방식으로 실행
- SearchService는 NaverSearchClient 의존 대신 SearchClientManager를 활용하도록 리팩토링
- PageResponseDTO 포맷은 그대로 유지하며 응답 통합
