+++ src/main/java/kr/hhplus/be/server/api/search/SearchController.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.api.search;
 import kr.hhplus.be.server.api.common.dto.ResponseDTO;
 import kr.hhplus.be.server.api.common.dto.PageResponseDTO;
 import kr.hhplus.be.server.api.search.dto.SearchResponseDTO;
-import kr.hhplus.be.server.application.SearchService;
+import kr.hhplus.be.server.application.search.service.SearchService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
